Fix for extra <cols/> element
The writer was inserting an orphaned <cols/> element in the worksheet xml whenever the column styles had not been specifically set. 
Line 307 in the original code is the culprit - even if no styles are set, a single <cols/> element will be returned, breaking the xml. 
Then, when you attempt to open the file in Excel, it says that errors were detected in the file.
For an example, the following code (taken from the docs) causes this error.
(write-xlsx
  "basic.xlsx"
  (lambda ()
    (add-data-sheet "Sheet1" '(("month1" "month2" "month3" "month4" "real")))
    (add-data-sheet "Sheet2" '((201601 100 110 1110 6.9)))))

My fix is simply to check before that there are column styles, and if not, (to-col-width-style) doesn't return anything.